ZMS/branches/zms4/requirements.txt
--index-url https://pypi.python.org/simple/

# fetch latest development snapshots from SVN/GIT-Repositories
# git+https://github.com/zopefoundation/Zope@2.13#egg=Zope2
# svn+https://zmslabs.org/svn/zmslabs/ZMS/trunk#egg=ZMS3
# git+https://bitbucket.org/zms3/behave.git#egg=zms3.behave
# git+https://bitbucket.org/zms3/formulator.git#egg=zms3.formulator
# git+https://bitbucket.org/zms3/mindmap.git#egg=zms3.mindmap
# git+https://bitbucket.org/zms3/photodb.git#egg=zms3.photodb
# git+https://bitbucket.org/zms3/qrcode.git#egg=zms3.qrcode
# git+https://bitbucket.org/zms3/themes.git#egg=zms3.themes

# fetch official releases from PyPI
AccessControl==3.0.12
Acquisition==4.2.2
BTrees==4.2.0
DateTime==4.1.1
DocumentTemplate==2.13.2
docutils==0.12
ExtensionClass==4.1.2
five.globalrequest==1.0
five.localsitemanager==2.0.5
initgroups==2.13.0
mechanize==0.2.5
Missing==3.1
MultiMapping==3.0
nt-svcutils==2.13.0
Persistence==2.13.2
persistent==4.2.0
Products.BTreeFolder2==2.14.0
Products.CMFCore==2.2.10
Products.ExternalMethod==2.13.1
Products.GenericSetup==1.8.3
Products.MailHost==2.13.2
Products.MIMETools==2.13.0
Products.OFSP==2.13.2
Products.PythonScripts==2.13.2
Products.StandardCacheManagers==2.13.1
Products.ZCatalog==3.1
Products.ZCTextIndex==2.13.5
pytz==2016.4
Record==3.1
RestrictedPython==3.6.0
six==1.10.0
tempstorage==3.0
ThreadLock==2.13.0
transaction==1.5.0
zc.lockfile==1.1.0
ZConfig==3.1.0
zdaemon==4.0.0
ZEO==4.2.0b1
zExceptions==3.0
zLOG==3.0
zodbpickle==0.6.0
ZODB==4.2.0
ZODB3==3.11.0
zope.annotation==4.4.1
zope.app.publication==3.14.0
zope.authentication==4.1.0
zope.browser==2.1.0
zope.browsermenu==4.1.0
zope.browserpage==4.1.0
zope.browserresource==4.1.0
zope.component==4.2.2
zope.configuration==4.0.3
zope.container==4.1.0
zope.contentprovider==4.0.0
zope.contenttype==4.1.0
zope.datetime==4.1.0
zope.deferredimport==4.1.0
zope.dottedname==4.1.0
zope.error==4.1.1
zope.event==4.2.0
zope.exceptions==4.0.8
zope.filerepresentation==4.1.0
zope.formlib==4.3.0
zope.globalrequest==1.0
zope.i18n==4.1.0
zope.i18nmessageid==4.0.3
zope.interface==4.1.3
zope.lifecycleevent==4.1.0
zope.location==4.0.3
zope.pagetemplate==4.2.1
zope.processlifetime==2.1.0
zope.proxy==4.2.0
zope.ptresource==4.0.0
zope.publisher==3.13.4
zope.schema==4.4.2
zope.security==4.0.3
zope.sendmail==3.7.5
zope.sequencesort==4.0.1
zope.site==4.0.0
zope.size==4.1.0
zope.structuredtext==4.1.0
zope.tal==4.2.0
zope.tales==4.1.1
zope.testbrowser==4.0.4
zope.testing==4.5.0
zope.traversing==4.0.0
zope.untrustedpython==4.0.0
zope.viewlet==4.0.0
Zope2==2.13.24
ZopeUndo==2.12.0
